[Minor 18] Prevent justification of table headings
Do this by using a new ragged2e package column type, and declaring affected
table columns as of this type.

Remove the p-value column of two tables, which do not add anything over
and above the q-value columns already present.

Also fix one capitalisation typo of 'treatment' in a heading